Changed int to double. Fixed bugs.

+++ ShoppingCartService/src/djangounchained/util/CartItem.java
@@ -6,7 +6,7 @@ public class CartItem {
 	//private String price = null;
 	private String description = null;
 	private int quantity = 0;
-	private int price = 0;
+	private double price = 0;
 	
 	
 	public int getQuantity() {
@@ -27,10 +27,10 @@ public class CartItem {
 	public void setName(String name) {
 		this.item_name = name;
 	}
-	public int getPrice() {
+	public double getPrice() {
 		return price;
 	}
-	public void setPrice(int price) {
+	public void setPrice(double price) {
 		this.price = price;
 	}
 	public String getDescription() {
